+++ cmd/ssql-playground/main.go
@@ -210,6 +210,7 @@ func buildCommand() *cf.Command {
 	cmd = commands.RegisterDescribe(cmd)
 	cmd = commands.RegisterUnpivot(cmd)
 	cmd = commands.RegisterFill(cmd)
+	cmd = commands.RegisterExtract(cmd)
 	cmd = commands.RegisterWhere(cmd)
 	cmd = commands.RegisterUpdate(cmd)
 	cmd = commands.RegisterCast(cmd)